# Reduce the number of branches in jit_exec